hotfix2: fix mistakes

+++ wallet-service/src/main/java/by/toukach/walletservice/controller/views/DebitAccountViewChain.java
@@ -7,6 +7,7 @@ import by.toukach.walletservice.enumiration.TransactionType;
 import by.toukach.walletservice.exception.ArgumentValueException;
 import by.toukach.walletservice.exception.EntityDuplicateException;
 import by.toukach.walletservice.exception.InsufficientFundsException;
+import java.math.BigDecimal;
 import java.util.Scanner;
 
 /**
@@ -24,7 +25,7 @@ public class DebitAccountViewChain extends TransactionViewChain {
     Scanner scanner = getScanner();
 
     System.out.println(ViewMessage.SUM);
-    double answer = scanner.nextDouble();
+    BigDecimal answer = scanner.nextBigDecimal();
     scanner.nextLine();
 
     TransactionDto transactionDto = TransactionDto.builder()
